rpc-client/pyproject.toml
[project]
name = "user-grpc-client"
version = "0.1.0"
description = "FastAPI client for User gRPC service"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "grpcio>=1.74.0",
    "grpcio-tools>=1.74.0",
    "uvicorn[standard]>=0.35.0",
    "pydantic>=2.11.6",
    "pydantic[email]>=2.11.6",
    "protobuf>=6.31.1",
    "opentelemetry-instrumentation-fastapi>=0.57b0",
    "opentelemetry-instrumentation-grpc>=0.57b0",
    "opentelemetry-instrumentation-logging>=0.57b0",
    "opentelemetry-sdk>=1.36.0",
    "opentelemetry-api>=1.36.0",
]

[tool.pyrefly]
exclude = ["proto/", "proto/*"]

[tool.pyrefly.errors]
missing-attribute = false
bad-argument-type = false

[tool.ruff]
exclude = ["proto/", "proto/*"]

[tool.ruff.lint]
exclude = ["proto/", "proto/*"]
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
